+++ src/Controller/PhotosAlbumsController.php
@@ -60,7 +60,11 @@ class PhotosAlbumsController extends AppController {
 			$this->redirect([$this->request->query('q')]);
 		
 		$this->set('album', $this->PhotosAlbums->find('active')
-			->contain(['Photos' => ['fields' => ['id', 'album_id', 'filename', 'description']]])
+			->contain(['Photos' => function($q) {
+				return $q
+					->select(['id', 'album_id', 'filename', 'description'])
+					->order([sprintf('%s.created', $this->PhotosAlbums->Photos->alias()) => 'DESC', sprintf('%s.id', $this->PhotosAlbums->Photos->alias()) => 'DESC']);
+			 }])
 			->select(['id', 'title'])
 			->where(compact('slug'))
 			->cache(sprintf('albums_view_%s', md5($slug)), $this->PhotosAlbums->cache)